requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.26
pyobjc-framework-Quartz>=10.0
